Add RefineCluster interface
